+++ merge.variants.sh
@@ -22,18 +22,25 @@ else
 	tool_info=$( cat $run_info | grep -w '^TOOL_INFO' | cut -d '=' -f2)	
 	script_path=$( cat $tool_info | grep -w '^WHOLEGENOME_PATH' | cut -d '=' -f2 )
 	SNV_caller=$( cat $run_info | grep -w '^SNV_CALLER' | cut -d '=' -f2 )
+	variant_type=$( cat $run_info | grep -w '^VARIANT_TYPE' | cut -d '=' -f2)
 	
-	ls $output_dir/*.chr${which_chr}.raw.snvs.bed.i.ToMerge | sort > $TempReports/chr${which_chr}.snvs
-	if [ $SNV_caller == "SNVmix" ]
-	then	
-		perl $script_path/snv.merger.pl -i $TempReports/chr${which_chr}.snvs -c probability -o $TempReports/list.chr${which_chr}.snvs
-	elif [ $SNV_caller == "GATK" ]
+	if [ $variant_type == "BOTH" -o $variant_type == "SNV" ]
 	then
-		perl $script_path/snv.merger.pl -i $TempReports/chr${which_chr}.snvs -c quality -o $TempReports/list.chr${which_chr}.snvs
+		ls $output_dir/*.chr${which_chr}.raw.snvs.bed.i.ToMerge | sort > $TempReports/chr${which_chr}.snvs
+		if [ $SNV_caller == "SNVmix" ]
+		then	
+			perl $script_path/snv.merger.pl -i $TempReports/chr${which_chr}.snvs -c probability -o $TempReports/list.chr${which_chr}.snvs
+		elif [ $SNV_caller == "GATK" ]
+		then
+			perl $script_path/snv.merger.pl -i $TempReports/chr${which_chr}.snvs -c quality -o $TempReports/list.chr${which_chr}.snvs
+		fi
+		rm $TempReports/chr${which_chr}.snvs
 	fi
-	rm $TempReports/chr${which_chr}.snvs
-	ls $output_dir/*.chr${which_chr}.raw.indels.bed.i.ToMerge | sort > $TempReports/chr${which_chr}.indels 
-	perl $script_path/indel.merger.pl -i $TempReports/chr${which_chr}.indels -o $TempReports/list.chr${which_chr}.indels
-	rm $TempReports/chr${which_chr}.indels
+	if [ $variant_type == "BOTH" -o $variant_type == "INDEL" ]
+	then
+		ls $output_dir/*.chr${which_chr}.raw.indels.bed.i.ToMerge | sort > $TempReports/chr${which_chr}.indels 
+		perl $script_path/indel.merger.pl -i $TempReports/chr${which_chr}.indels -o $TempReports/list.chr${which_chr}.indels
+		rm $TempReports/chr${which_chr}.indels
+	fi
 	echo `date`
 fi
